libflux/reduce: rework API and fix bugs
Change flux_red_t to flux_reduce_t and make it refer to an incomplete
struct rather than a pointer to one (RFC 7).

Rename FLUX_RED_* flags to FLUX_REDUCE_*.

Instead of individual setters for reduction ops, create a structure
of function pointers that is passed into flux_reduce_create().

Make reduce and sink ops have the same prototype.

Add destroy op so that items can be disposed of if sink is not defined.

Add forward op.  Sink used to do double duty as sink and forward
and was always implemented as a conditional with two branches.
Ignore "forward" on rank 0, and "sink" on rank != 0.

Drop flux_reduce_flush() from public API.

Drop flux_redstack_t type and add functions that operate directly
on the flux_reduce_t type: flux_reduce_push(), flux_reduce_pop().

Added itemweight op that allows an item to be interrogated
as to how many original items it represents.  If using the HWM
flush policy, this is required (or EINVAL).

Only call reduce if >1 items are queued

Change timeout to seconds and make the type a double.
Stagger the timeouts based on the level in the tree,
scaled down at the leaves, full timeout at the root.
Since there may be multiple items to reduce at the leaves,
avoid setting TIMEDFLUSH timer there to zero.  Use the tree
height plus one to calculate it.

Ensure that the reactor timer is reset between calls.

Add flux_reduce_opt_get/set() for accessing
- learned high water mark
- calculated timeout values
- unweighted item count
- weighted item count