binding/python: cleanup before dist
Since python 'test' and 'test_runner' directories are included
recursively in the EXTRA_DIST target, you get *everything* in the
dist tarball. This isn't a problem with a clean working directory,
but if you are doing 'make dist' or 'make distcheck' from a working
direcory you may get unwanted files in the distribution like .pyc
files and .log and .trs from the testsuite.

Automake provides the 'dist-hook' target for just this purpose,
so this patch uses that hook to do a little cleaning before dist.

This actually fixes a 'distcheck' failure if the make check
test_runner.log file is in the working tree before the command
is run.