connectors/local: Re-connect loop non-infinite
When attempting to connect to a local flux instance, the loop previously only
checked for the existence of the socket file and that a live matching pid must
exist.  In the, now less likely but still possible, event that a new process
other than the flux instance that created the socket inherits the pid, the
loop would hang indefinitely.  This commit adds an upper limit, currently set
to 5.