test: add loop test for reduction
The loop test is able to exercise all the flush policies without
actually instantiating a flux session.  It does this by creating
a single flux_reduce_t queue, putting items in it, and testing which
callbacks are made.  For TIMEDWAIT policy, it starts a reactor loop
so that timers can fire.

Edge cases such as tardy items from earlier batches are covered.