cmd: prepend flux cmddriver dir to PATH
If the flux(1) command driver is not run as literally "flux"
prepend the directory path of the current executable to PATH.

This allows processes invoked from flux(1) command to find
specifically *this* flux driver first in PATH. Therefore things
like

 src/cmd/flux start --size=2 flux exec hostname

will work as expected (i.e. the correct flux command driver is
called.)

If the command driver was invoked as "flux" then it can
be assumed that the correct flux will be found and the PATH is
not modified.

Because we now reuse the `dir_self` function, the results are
memoized, and `.libs` is always stripped from the end of the result
to handle use in a build tree.

Fixes #263.