libutil: adding Redis string, sds, library and vec
New support libraries for easier manipulation of dnamic-sized data, pulled in
to ease some slightly complicated string management for hwloc support and
more precise environment handling.  While that is the primary reason, sds is
also fast, binary compatible, and very good for slicing and dicing partially
send messages and so-forth, which is what Redis uses it for.  The vec library
is just a simple way to dynamically generate dense arrays, which makes working
with the split and join functions from sds, as well as anything that requires
a dynamic argv-style input, easier.