adds code formatters for c and python
The .clang-format file defines a style following the rules set down in our RFC
on the subject, and enforcing them across an entire file.  If possible, the
lines at the bottom should be uncommented to get the best alignment results,
but this requires a newer clang than we have on LC machines by default.  The
file as committed is compatible with the default clang 3.5 that is generally
available.  Use it just by running clang-format on a file, or a piped input
stream, from any directory below the file, it will be found and used as the
configuration automatically.  If possible, it would be nice to use this by
default, but for a while it will cause ancillary changes on pull-requests to
do that unless we do a global re-format at the start.

The .style.yapf file is an equivalent system for python, it requires the yapf
(Yet Another Python Formatter) program to be available, which is easily
installed via pip or spack.  It works the same way that clang-format does for
the C code, and enforces a PEP8-compatible style that uses the same spacing
and general rules as our C style where possible, and as close as feasible
where the rules conflict with python language requirements.