conf/env: functions renamed and lua path reworked
The "split and push while de-duplicating" functions are now named
`*_upsert_{front,back}` to reflect their alternative behavior.  If
anyone has other ideas for this, let me know, this was the closest I
could seem to come. This commit also adds the push functions back as
explicit non-deduplicating push functions, and uses one of them to
simplify the handling of the lua default path.