config/env: properly handle contiguous separators
This tests the complex path de-duplication, where an incoming path
already contains duplicates, as well as preservation of a user's
external environment value.

After the switch in lua path handling, empty path components returned by
split were not being handled correctly, causing an extra path separator
to be added the first time it was encountered.  This commit adds a test
for this case, and a check to ensure that empty components don't add
extra separators.

The names should be a bit more intuitive for new users, also the interface
functions have been documented in conf.h with doxygen compliant doc-comments.