broker/snoop: avoid message copy, cleanup
When the on-demand snoop socket is up and running, avoid copying
each message by switching from zmsg_send() to flux_msg_sendzsock().
Change the prototype of snoop_sendmsg() to accept a
(const flux_msg_t *).

If snoop_set_sec() is not called, don't try to initialize security
and trigger an assertion in czmq (needed for testing).

Add some explanatory docs in the header file.