python/kvs: use kvsdir size and increase lazyness
Now that a kvsdir length function is available, the python bindings should
take advantage of it.  In connection with that, it's now relatively efficient
to greate generators for walk instead of lists, which should make walks vastly
more efficient in the general case where only some of the data gets used.