cmd/flux-start: update broker options, drop slurm
Drop slurm launch method now that the broker can be launched
directly with no arguments.

Call with --boot-method=LOCAL for direct launch.

Create the local socket directory and pass it on the command
line to all the brokers so tey can share it yet have it
generated uniquely.

Append the initial program command and args to the broker
command line instead of --command.