zio: don't reschedule zio writer if dstfd closed
A zio writer, such as stdin, could be scheduled on the reactor loop
even after the dstfd was closed, because zio_write_pending() wasn't
checking if the zio writer object was already "closed". In this
context, a zio object is closed if EOF has been sent, but for a dstfd,
the EOF_SENT flag was not being set.

This patch sets the flag when close(2) is called on the zio writer's
dstfd. This not only fixes busy loop possibility in the reactor loop
(since zio writer callback will constantly wake up), but also fixes
a case where a zio callback could return -1 and exit early from the
reactor, which may cause all kinds of problems in users like wrexecd,
which exits if the flux reactor terminates for any reason.

This may be a cause of wreck hangs seen in #148 and #357.