broker: allow flux to launch flux
Add a config file entry for flux_program_libpath which points
to the set of directories containing shared libraries, notably
libpmi.so and libflux-core.so, corresponding to the current instance.

Set LD_LIBRARY_PATH to above:$(fluxlibdir):$LD_LIBRARY_PATH,
taking care to dedup.  FIXME: this code is not as robust or tested
as the path dedep code in libflux/conf.c.  Consider how to make
this code work with subprocess environment.

Change the logic to always execute the initial program on rank 0,
unless the --noshell option is provided.