wreck: support events over local connector in wrexecd
Support events (just the "kill" or signal event for now) over the
already existing local flux connector in wrexecd. This will eventually
replace the custom and unnecessary dealer/router socket connecting
wrexecd to the wrexec module in the enclosing flux-broker.

In order to avoid all wrexecd processes listening to all wrexec
events, we tweak the supported signal event to have the jobid
in the topic string. So, the topic becomes "wreck.<id>.kill" instead
of "wreck.kill".