local connector: rework subscription management
The local connector implements subscribe/unsubscribe ops as requests
to the connector-local module.  Replace the fire and forget requests
with an RPC.  This fixes #375.

The local connector acts as an event re-publisher, managing its own
list of subscriptions.  We were maintaining subscription topics in
a per-client list.  Subscribing consisted of adding a list entry
and sending a cmb.sub request to the broker.  Unsubscribing consisted
of deleting a list entry and sending a cmb.unsub request to the broker.
Re-publishing an event consisted of walking the list of clients and
each client's list of topics.

Replace this with a global and per client hashes of subscription topics,
with use counts.  This allows round trips between connector-local and
the broker to be minimized when multiple clients subscribe to the same
event.  It also allows a linear search in each client to be replaced
with a hash lookup in the event re-publishing path, at least when event
topic string exactly matches the subscription.

The replaced code had this bug, now squashed:  when unsubscribing (as
opposed to disconnecting), we removed the subscription from the local
list but failed to inform broker or free the storage.