testsuite: fix wreck signal test
The test for signal forwarding in wreckrun was broken. For some reason
executing under 'run_timeout' was breaking signal delivery or default
signal handling for flux-wreckrun.

Instead of run_timeout use a shorter sleep to ensure we don't wait
for long if test fails, and run flux-wreckrun outside of any timeout
wrapper.