connectors/loop: don't add route stack
Don't push the uuid on requests in the loop connector, as
responses will have an extra hop in their route stack, which
results in test failures after the tightening up of flux_msg_match().

There is no reason to append a route stack to events either.
They are not going anywhere.