connectors/local: require path to be set
In the connector itself, don't accept a NULL path.
The full URI must be provided.

In modules/connector-local, obtain the URI of the local socket
by asking the broker with flux_getattr ("local-uri") instead of
deriving it from FLUX_TMPDIR.

Rename the socket from flux-api to local in both places (cleanup).