broker: epgm turned off by default
Instead of always enabling epgm when boostrapped via PMI, only
enable it when specified on the command line, e.g.
  flux start -o,--enable-pgm ...

This makes it possible for flux to launch flux on systems (such
as test systems) on which epgm may not work.  It also avoids
several other epgm related issues discussed in #385.

If epgm is not enabled, events are propagated via the TBON.
This has some disadvantages:
* higher latency and more of a ripple effect
* TBON self-healing algorithm depends on a separate channel for events
These might be acceptable tradeoffs for small instances, especially
of one or a few nodes.