broker: drop -c,--command
Drop the -c,--command option and instead aggregate all args
at the end of the command line into the initial program command
and arguments.  Hopefully this will avoid the obligatory quoting
for multi-word commands, which an be tricky in shell scripts, especially
when launching flux recursively.

Make getopt options posixly-correct.