broker/attr: add attr class for broker use
The attribute class used by the broker is a bit more sophisticated
than the one offered to users.

The broker has the ability to set flags on attributes, to influence
how they are cached handled by users and to limit users' ability
to change them.

"active attributes" are supported.  These attributes have get/put
callbacks associated with them so that when a user calls getattr,
a function is called in the broker to compute the value; or when a
user calls setattr, a function is called to take some action in the broker.

The active interface also provides convenience functions to "watch"
integer values as attributes.  When a user calls getattr, an internal
callback reads the integer value and converts it to a string value
and for setattr the reverse.