libflux/flog: add flux_log_error, flux_log_verror
These functions log messages to LOG_ERR level, appending a colon,
space, and the string representation of errno.

This can replace the many places where we call
  flux_log (h, LOG_ERR, "descriptive message: %s", strerror (errno));
with the terse
  flux_log_error (h, "descriptive message");

Moreover, zmq_strerror() is used so that both system and zeromq
errno values are decoded, as was done in libutil/log.c.

The FLUX_LOG_ERROR macro logs file::line[function] and errno string.
It can replace the many places where we call (for unlikely errors):
    flux_log (h, LOG_ERR, "%s: %s", __FUNCTION__, strerror (errno));
with the terse
    FLOG_LOG_ERROR (h);

As a bonus, file::line will make it easier to find the error source.