broker/log: add new broker log class
Add a broker class for handling log messages generated
within the broker or received on the broker's cmb.log service.

All log messages of the local rank are stored in a ring buffer for
access via cmb.dmesg and cmb.dmesg.clear RPC's.

Each log entry stored in the ring buffer is given a monotonically
increasing sequence number.  cmb.dmesg provides a 'seq' argument,
-1 on the first call, equal to the previous returned log entry's
sequence number on subsequent calls.  The next largrest sequence entry
is returned, ENOENT if there are none.  cmb.dmesg.clear provides a
'seq' argument used to clear messages with sequence number <= seq.

Logs at level <= log-level are also routed upstream from rank > 0,
and to a configured log destination on rank 0.

Broker command line options:
   -l,--log-level=N   - level of logging to be captured at rank 0
   -L,--logdest=FILE  - where rank 0 should store logs (default stderr)

Changes to -L,--logdest option argument interpretation:
- the file named "syslog[:facility]" is no longer interpreted
- the file named "stdout" is no longer interpreted
- the file named "stderr" is special, as before (sideflux.py uses this)

The following attributes were added:
   log-level (rw)     - level of logging to be captured at rank 0
   log-buflimit (rw)  - size of ring buffer
   log-bufcount (r)   - count of entries in ring buffer
   log-count (r)      - total count of log messages (including wrapped)

A "nonblock" flag is supported in the dmesg request such that if the
flag is set and the dmesg call would return ENOENT, push the request onto
a list and restart the request handler when a message is added to the log.