modules/wreck: use flux_log_error() for waitpid
wrexec.c was calling err() on waitpid() error.  This was coming
out on stderr of the t2000-wreck.t after the broker -L option was
made to apply only to log output.

Use flux_log_error() instead.

Possibly this error is not really an error and the message
should be completely suppressed or logged at debug level.