build/coverage: don't report python ffi coverage
Something seems to be wrong with coverage testing of the generated
C bindings for python. A lot of python tests are run, but the coverage
is only reporting at most a couple lines. Possibly something about
the way python C bindings are used is confusing gcov, but I'm not
smart enough to fix it, so just ignore python bindings in lcov
output for now.