libflux/reactor: unify watchers in flux_watcher_t
Create a generic watcher type flux_watcher_t for all watchers.
Instead of having a separate type, callback prototype, create,
destroy, start, and stop for each kind of watcher, only a separate
create function is needed for each kind, and possibly an accessor
to compensate for the single callback prototype.