lua: kvs: do not convert string to number on put
During a kvsdir newindex method (i.e. kvsdir_put()) the lua bindings
were inadvertently converting string values to numeric by using
lua_isnumber (L, index), which, if successful coverts a string to a
number in the stack.

Use lua_type (L, index) == LUA_TNUMBER instead, so that a string
assigned to a kvsdir object remains a string in the kvs.

This coercion is the documented behavior of lua_isnumber(), though
easy to forget:

 http://pgl.yoyo.org/luai/i/lua_isnumber