wreck: Fix globals in luastack code
Scripts loaded by luastack module were supposed to be sandboxed such
that globals for each script were in their own namespace. This was
not working as designed and was broken when loading more than one script.

Fix code the line that overrides the lua globals index for these scripts.

Note: this temporarily breaks Lua 5.2 compatibility.