libflux/dispatch: add flux_msg_handler_t
Since the "message watcher" isn't really a reactor watcher, create
a new object called a "msg_handler" with its own create/start/stop/destroy
methods distinct from the reactor.

The reactor is used indirectly, since a handle watcher is implicitily
registered/unregistered as handlers are registered/unregistered.
flux_get_reactor() is used to obtain the "default" reactor for
the handle.  This is created on demand, or can be overridden by
calling flux_set_reactor() before the first handler is registered.

Update usage of the coproc API to reflect RFC 7 changes.