test/coproc: don't emit extra test result on fail
The final coproc test tries to elicit a segfault by scribbling
on a coproc stack's guard page.  When this doesn't trigger a
segfault, the coproc's callback was emitting an extra "success"
result, in addition to the failure.  Fix the test so that the
coproc callback only emits a diag message.

Also make some of the test output a bit more informative.

Partially fixes #415