travis-ci: replace openmpi with mpich
Incorporate mpich into the travis-ci environment, so we can test
simple-v1 PMI support, and also Flux being launched by mpiexec/hydra.

Unfortunately we cannot apt-get the mpich package because it has not
been whitelisted for use within travis, so fetch and build it with
travis-dep-builder.

There is a known issue with clang segfaulting while building mpich.
Work around this by setting CC=gcc in travis-dep-builder, so that
dependent packages are built using gcc in both gcc and clang builders.