bindings/lua: fix crash in lua_push_flux_handle
Under at least one version of Lua 5.1 an assertion was being
triggered in lua_push_flux_handle() when under use by the lua
plugins in wrexecd. The invariant that was violated was that
the "flux" entry existed in the flux reftable, but the userdata
pointer to the real flux_t handle was invalid.

It is unclear why this was happening, but may be due to the
local/global/coroutine shenanigans used in luastack.c to
run multiple plugins against the same global state.

In this case, it is safe to recreate the flux_t userdata, so instead
of asserting here just fall through if no userdata found.

Fixes #423.