libpmi-client: add pmi client library
Flux will use this library to bootstrap from a PM, including itself.
We will also use this library to test our PMI implementation.

Initially there are two embeded implementations:

"dlopen" - this implementation dlopens libpmi.so.  Functions that
are undefined will fail gracefully wtih a PMI_FAIL return code.

"simple" - this implements the simple v1 PMI wire protocol used
by mpich and hydra.  It is sufficient to allow hydra to launch Flux,
but otherwise is very rigid about what it expects on the wire,
and is missing a few functions that could be implemented.

The pmi client object can be initialized to explicitly use one of
these implementations with parameters such as a fully qualified path
to a library to dlopen, or pmi_create_guess() can be called to
pick one based on the environment.