cmd/flux-zio: strip down and rename to t/kz/kzutil
The flux-zio program is no longer the only way to peek at a kz
stream created by wreckrun, and it no longer has much to do with zio,
so let's move it to t/kz/kzutil, to be used by a sharness kz test driver.

Drop the --run option (and forkzio class).
Rename --raw to --raw-tty.
Rename --force to --trunc, and --lazy to --delay-commit to reflect
KZ_FLAGS_TRUNC and KZ_FLAGS_DELAYCOMMIT, respectively.

Modernize reactor usages and tidy up.