wreck: abstract job io into wreck.io module
Add src/bindings/lua/wreck/io.lua module with abstraction
for job/task output (output only for now, stdin will come later)
so that wreck-based job output can be managed from multiple
tools and utilities.

The abstraction presents as a simple io multiplexer that funnels
the many task io streams to one or more destination streams.
The streams are reference counted so that io "completion" can
easily be tested, and the module allows a completion handler
to be registered as a callback when all IO is complete.