libutil: optparse: support repeated option usage
Change internal optparse_t option "optarg" field to an "optargs"
list and save all option arguments when an option is used multiple
times. (Though for now we still return the last optarg given on cmdline
by default).

Promote opt->found to an integer from a single bit and use it to record
the number of times an option was seen. Now optparse_getopt (p, "name")
returns the total number of times option "name" was used.