modules/libsubprocess: use child watchers
If a reactor has been set in the subprocess manager, start a
child watcher on it for each new subprocess.  Then when the process
exits, use the child watcher callback to set its status and call
the exit callback, if any.

It's possible to start some subprocesses, start the reactor,
and expect the reactor to terminate once all the children (and any
active zio's) have terminated/stoppped.

If the reactor was not created with FLUX_REACTOR_SIGCHLD,
creation of the child watcher will fail.  Ignore that error
It just means that subprocess_reap_all() will have to be called
to obtain status.

Add a new status callback that (if registered) is called whenever a
process changes state, including the non-terminating state changes of
starting and stopping.  This is in addition to the exit callback which
probably should remain as is, since it is called when both the process has
been reaped, and I/O has been completed.

  int subprocess_set_status_callback (struct subprocess *p,
                                      subprocess_cb_f fn, void *arg)

Add accessors for stopped/continued state:

  int subprocess_stopped (struct subprocess *p)
  int subprocess_continued (struct subprocess *p)