wreck/lua: Initial version of stdin handling plugin
New input.lua plugin handles stdin configuration registered as a table
in lwj.input.config for all jobs or lwj.<id>.input.config for the
current job. The input config is a list of src/dst pairs, where src
is a file or "stdin" for the standard stdin stream of a job (located at
lwj.<id>.input.files.stdin), and dst is a list of task ids to link
to this stream, or "all" or "*" for all tasks.

The plugin works by opening each stream once as a file in the kvs under
lwj.<id>.input.files.<n>.kz, then using symlinks to direct requested
taskid stdin to these files. Thus an input stream need only be written
once to the kvs to be read for all destination tasks.