wreck/lua: add -i, --input option
Add -i, --input option to set input configuration as expected
by the input handling plugin. The --input option takes a
semicolon separated list of the form

  [src:]dst[;...]

where src is a source file or "-" or "stdin" to indicate the
standard input stream (from a front end command like wreckrun)
and dst is a list of tasks or "all" or "*".

E.g. to link stdin to task 0:      --input=0
     Explicitly bcast stdin:       --input=-:*
     stdin to task 0, /tmp/foo to all other tasks:
                                   --input=/tmp/foo:*
     Don't do anything with stdin: --input=none