cmd/flux-start: use subprocess status callback
Use the subprocess status callback to display all process state
changes on stderr.  Use the exit callback to start/stop the exit
timeout and destroy subprocess state.

Don't exclude SIGTERM from the signals that contribute to a failed
exti code.  If the session is stuck and cannot exit via the broker
shutdown timer before the exit timeout, then something is wrong that
we need to know about.