broker: block signals early
Set the signal mask to block all signals before initializing zeromq
or starting module threads, so that all threads will inherit the fully
blocked sigmask.  Both actually do this explicitly anyway, but allowing
any signals to be set at pthread_create time opens a window where a
signal could be misdirected.

Move zsys_handler (NULL) to right before the zctx_new() call.
This is where it needs to be to prevent czmq's default signal handlers
from being installed.  (It was OK where it was but this reduces the
likelihood that it will be placed out of order later).