broker: add --shared-ipc-namespace option
If the broker is started with --shared-ipc-namespace, it will create its
TBON request sockets in the ipc:// space rather than tcp://.
This should only be used if it is known that all brokers are co-located
on the same node and share the same $TMPDIR namespace.

Also, this option directs the broker to use a single ipc socket for
event distribution if epgm is disabled, rather than piggy backing it
over the TBON.

flux-start uses this to avoid claiming tcp ports unncessarily in the
test environment.