alternate hwloc support
This branch adds new configuration options to specify hwloc xml files to
load in place of dynamically determining the system topology.  At
present the files must be visible to the ranks they are specified for,
and the default must be accessible from all ranks.  This restriction
could be lifted later.