t/t2001-jsc.t: reduce time spent sleeping
Sleep 1/4 of the time in overlapping wreckrun tests. This should
be sufficient to get the tests to overlap, but not waste time waiting
for longest sleep to complete.

Also, instead of sending flux-wreckrun into background, use --detach
to detach immediately after issuing runrequest. This lets us collect
jobids and wait for each job with `flux wreck attach`.