build: define fluxlib_ldflags, fluxmod_ldflags
Instead of repeating ldflags throughout the src/modules/*/Makefile.am,
define a set of common flags in configure.ac.

Note the -module option is parsed by libtool and becomes invisible
if defined inside an AC_SUBST, so that one has to be specified in
each Makefile.am beside $(fluxmod_ldflags) or libtool complains that
comms modules don't have a "lib" prefix.

Also note that the default for $(fluxlib_ldflags) is to export
all symbols, so it should be paired wtih a -export-symbol-regex
option specific to the library.

Update src/modules/*/Makefile.am to use the new macros.
(Also the couple modules located in src/test/*)