modules/wreck: fix wrexecd initialization failures
In rexecd_init() errnum was not manually initialized to zero, so in
some cases (where randomly errno gets initialized to value other than
zero), flux wreckrun could fail incorrectly "at initialization" since
the value of errnum is used to determine a global failure.

This fix initializes errnum to zero, and sets it to 1 on negative
return code from flux_kvs_get() when checking for "fatalerror" in
kvs.