libflux/rpc: add raw payload functions
As described in the previous commit, Flux messages per
RFC 3 allow JSON or "raw" message payloads in addition to
JSON payloads.  This is useful in some cases where JSON encoding
is an unnecessary hinderance, such as when sending and receiving
binary blobs.

This PR adds variants of flux_rpc() and flux_rpc_get() that
encode/decode request/response messages with raw payloads.
The functions behave identically to their JSON counterparts
and may be intermixed for a given RPC.