modules/wreck: only support job.submit if sched is loaded
Only support job.submit service if a "sched" module is loaded.
The existence of sched is determined by attempting a ping request to
module (rpc to sched.ping service).

In case sched module is loaded after the fact, a sched.ping is attempted
on each job.submit request until successful, after which it is assumed
the sched module will stay loaded, so not further pings are made.