modules/wreck: fix race in signal delivery
If killpg(2) fails during prog_ctx_signal() it is likely that we're
in the window where the child has not yet been assigned a new process
group id, so try killing the task pid directly with kill. Since setpgrp()
is called before the task would have any chance to invoke further
children, this should not result in creation of orphans.

Fixes #448.