modules/content-sqlite: sqlite backing store
Provide the 'content-backing' service using a sqlite.
It is unoptimized; there is no batching of commits yet.

Notify the content-cache that a content service is available
when this module is loaded using 'content.persist' RPC.
At module unload time, notify 'content' that the content
service is going away.

Notice when the broker is shutting down and disable the transfer
of data to the cache at unload (although at the moment the
broker does not attempt to unload modules at shutdown).