python: add minimal python kz support
This patch adds a kz python module with the ability to "attach" a kz
stream to an output stream or file descriptor.  No support for arbitrary
interactions with a kz stream as yet, but sufficient to support cap/ATS
efforts for the time being.

In creating this, I found that python cffi does not like unnamed enums,
hence the name added to the kz.h header's enum, and accessing those
enums exposed a bug in the wrapper generator's accesses of C-defined
non-callable, non-define data structures.