libutil/sha1: add sha1 from ghostscript
Pull in a well-tested SHA1 implementation as an alternative to
czmq's zdigest.  The problem with zdigest is that it requires
a small malloc to create the zdigest object each time it is used.
The content cache/service will compute a lot of digests and the
extra malloc overhead should be avoided if possible.