broker/content-cache: new service
Add a broker-resident service that provides a content-addressable
storage abstraction with distributed cache.  This is a rewrite of the
KVS distributed object cache, built for opaque content, optional
persistence, and operation independent from the KVS.

The cache provides the same semantics and scalability properties as the
KVS object cache.  That is, a store on any rank returns only when the
content can be loaded from any other rank.  Stores of identical content
are "squashed".  Loads of identical content benefit from caching at each
tree level.

The content-cache on rank 0 optionally uses a 'content' service for
backing store.  Use is not attempted until a 'content-store.persist'
request is made by a module providing the content service at module
load time.  Once enabled, stores on rank 0 are "write back" with respect
to the backing store, allowing "write through" store requests to proceed
at the speed of the content cache while content is copied to the backing
store asynchronously.  Once stored persistently, rank 0 cache entries can
be aged out of the cache as on other ranks.

This builtin service uses a new method to register handlers within
the broker.  A broker service handler for the 'content-cache' is
registered, which simply requeues request messages for that service
on the broker flux_t handle.  Message handlers for the service's
individual methods then be registered as flux_msg_handler_t's.

Each heartbeat, the cache is walked to find entries that want
expiring, until hardwired size (sum of all entries' sizes) and
numbert of entries targets are reached.