cmd/flux: no fatal error with -F and no conffile
It is not clear if use of `flux -F` with missing config file
should be a fatal error or not. However, there are cases where
use of -F is required (to force FLUX_URI to be ignored for example),
but there is no config file present or necessary (for installed
flux for example).

Therefore, as at least a temporary work-around, this patch updates
flux.c config file handling to *allow* -F with missing config file.