t/t0001-basic: robustly handle flux-config tests
When running under installed flux(1), KVS config may not be present
and is not required for Flux operation. The flux-config tests in
t0001-basic.t made an assumption that general.broker_path would
exist in the config. This patch removes that assumption, and handle
both cases where general.broker_path exists and when it does not.