libflux/request, response: allow payload size=0
Change the raw request/response encoding/decoding to not
return EPROTO if a payload is expected and none is returned,
and vice-versa.  This in turn affects the way the raw rpc
handle payloads.

This was needed to cleanly handle load/store of a zero size
blob in the content-cache.  Since the content-cache is the
only user of the raw rpc functions thus far, this seemed OK.

Update tests.
Update man pages.