modules/kvs: integrate content store
Drop the KVS "load" and "store" operations, which were used to propagate
object cache entries up and down the TBON.  Instead, use the content
service, and its distributed "blob" cache.  This communication between KVS
and content service is local (same rank).

The KVS continues to cache JSON objects (decoded blobs).  A future PR
will try to pare down the lifetime of entries in the object cache as
the content store is now caching the same data locally.

Since the responsibility for long-term retention of blobs now belongs
to the content service, the rank 0 KVS object cache can expire entries
as on other ranks.

Side note:  when storing JSON values in the content store, store the
terminating NULL so that the incremental parser knows it has ingested
a complete value.  This arises when bare integers are stored, for example.