modules/kvs: simplify commit_unroll()
commit_unroll() was designed with incorrect assumptions about what
is legal to do to a json_object while iterating over its values.
According to comments in json_object.h:

  Adding keys to the object while iterating is NOT allowed.

  Deleting an existing key, or replacing an existing key with a
  new value IS allowed.

Thus it is OK to call json_object_object_add() to replace the value
of an existing key while iterating over the object containing it,
and unnecessary to collect these changes in a separate hash, and apply
them afterwards.

This change reduces heap activity during commit processing on the
rank 0 kvs instance.