broker: new attr 'content-blob-size-limit'
Instead of hardwiring the content blob size limit per RFC 10,
make it a broker attribute so that it can be configurable.
It is set with the "immutable" flag so we do not have to make
it coherent across all users at runtime.  It thus can only be
changed at compile time.

We need this on a temporary basis while the KVS is in transition
to the "treeobj" object representation described in RFC 11.